+++ crates/gitbutler-oplog/Cargo.toml
@@ -15,6 +15,7 @@ strum = { version = "0.26", features = ["derive"] }
 tracing.workspace = true
 gix = { workspace = true, features = ["dirwalk", "credentials", "parallel"] }
 toml.workspace = true
+gitbutler-command-context.workspace = true
 gitbutler-project.workspace = true
 gitbutler-branch.workspace = true
 gitbutler-serde.workspace = true
